pyproject.toml
[tool.poetry]
name = "nnarch"
version = "0.1.0"
description = "Model implementations from first principles in python"
authors = ["Hannes Kimara <hanneskimara@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
click = "^8.1.7"
tqdm = "^4.66.2"
requests = "^2.31.0"
pillow = "^10.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
